Refactor updater runtime and transport boundaries
Split updater runtime assembly into core, release, workflow, and transport bundles.
Move transport-neutral uplink readiness into the transport package and make cleanup error context explicit in workflow and transport finalization.
