+++ apps/server/tests/use_cases/updates/wifi/test_wifi_readiness.py
@@ -2,16 +2,15 @@ from __future__ import annotations
 
 from dataclasses import replace
 from pathlib import Path
-from unittest.mock import AsyncMock
 
 import pytest
 from test_support.update_status import build_update_status_harness
 from use_cases.updates._update_manager_test_helpers import FakeRunner
 
 from vibesensor.use_cases.updates.runner import UpdateCommandExecutor
-from vibesensor.use_cases.updates.transport_failures import UpdateTransportStepError
+from vibesensor.use_cases.updates.transport.failures import UpdateTransportStepError
+from vibesensor.use_cases.updates.transport.uplink_readiness import UpdateUplinkReadiness
 from vibesensor.use_cases.updates.wifi.wifi_config import build_default_wifi_config
-from vibesensor.use_cases.updates.wifi.wifi_readiness import UpdateWifiReadiness
 
 
 def _build_readiness(
@@ -19,7 +18,7 @@ def _build_readiness(
     *,
     dns_ready_min_wait_s: float = 0.05,
     dns_retry_interval_s: float = 0.01,
-) -> tuple[UpdateWifiReadiness, FakeRunner]:
+) -> tuple[UpdateUplinkReadiness, FakeRunner]:
     runner = FakeRunner()
     status = build_update_status_harness(tmp_path / "state.json")
     config = replace(
@@ -28,39 +27,12 @@ def _build_readiness(
         dns_retry_interval_s=dns_retry_interval_s,
     )
     commands = UpdateCommandExecutor(runner=runner)
-    readiness = UpdateWifiReadiness(
+    readiness = UpdateUplinkReadiness(
         commands=commands,
         status=status,
         config=config,
     )
     return readiness, runner
-
-
-@pytest.mark.asyncio
-async def test_bring_uplink_up_retries_ssid_not_found_then_succeeds(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    readiness, runner = _build_readiness(tmp_path)
-    original_run = runner.run
-    connect_calls = {"count": 0}
-
-    async def run_with_retry(args, *, timeout=30, env=None):
-        joined = " ".join(args)
-        if "connection up VibeSensor-Uplink" in joined:
-            connect_calls["count"] += 1
-            if connect_calls["count"] == 1:
-                return (10, "", "Error: No network with SSID 'TestNet' found.\n")
-        return await original_run(args, timeout=timeout, env=env)
-
-    runner.run = run_with_retry
-    sleep = AsyncMock(return_value=None)
-    monkeypatch.setattr("vibesensor.use_cases.updates.wifi.wifi_readiness.asyncio.sleep", sleep)
-
-    await readiness.bring_uplink_up("TestNet")
-    assert connect_calls["count"] == 2
-    assert sleep.await_count == 1
-    assert any("dev wifi list ifname wlan0" in " ".join(call[0]) for call in runner.calls)
 
 
 @pytest.mark.asyncio
