+++ apps/server/vibesensor/use_cases/updates/runtime_core.py
@@ -1,0 +1,57 @@
+"""Core updater runtime assembly for status tracking and command execution."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from pathlib import Path
+
+from vibesensor.use_cases.updates.runner import (
+    CommandRunner,
+    UpdateCommandExecutor,
+    UpdateStatusCommandReporter,
+)
+from vibesensor.use_cases.updates.status import (
+    UpdateStateStore,
+    UpdateStatusTracker,
+    build_update_status_tracker,
+    collect_runtime_details,
+)
+
+__all__ = ["UpdateRuntimeCore", "build_update_runtime_core"]
+
+
+@dataclass(frozen=True, slots=True)
+class UpdateRuntimeCore:
+    status: UpdateStatusTracker
+    commands: UpdateCommandExecutor
+
+
+def build_update_runtime_core(
+    *,
+    runner: CommandRunner,
+    repo: Path,
+    state_store: UpdateStateStore,
+) -> UpdateRuntimeCore:
+    status = _build_status_tracker(
+        repo=repo,
+        state_store=state_store,
+    )
+    commands = UpdateCommandExecutor(
+        runner=runner,
+        reporter=UpdateStatusCommandReporter(status=status),
+    )
+    return UpdateRuntimeCore(status=status, commands=commands)
+
+
+def _build_status_tracker(
+    *,
+    repo: Path,
+    state_store: UpdateStateStore,
+) -> UpdateStatusTracker:
+    loaded = state_store.load()
+    status = build_update_status_tracker(
+        state_store=state_store,
+        status=loaded,
+    )
+    status.set_runtime(collect_runtime_details(repo))
+    return status
