+++ apps/server/vibesensor/use_cases/updates/transport/__init__.py
@@ -1,0 +1,1 @@
+"""Transport-focused updater boundaries and implementations."""
